Moved the logic of pagination to note-list-table component - Fixed indentation. - Added logic for the pagination count. - Moved the logic for pagination to note-list-table component. - Updated header design.
