+++ time-app/layouts/default.vue
@@ -1,4 +1,3 @@
-<script lang="ts" setup></script>
 <template>
   <div>
     <div class="p-4 font-bold shadow-md bg-blue-50 ring-2 ring-gray-200">
