+++ time-app/nuxt.config.ts
@@ -1,8 +1,8 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   devtools: { enabled: true },
-  modules: ['@nuxtjs/tailwindcss'],
-  build:{
-    transpile: ['vue-toastification'],
-  }
-})
+  modules: ["@nuxtjs/tailwindcss"],
+  build: {
+    transpile: ["vue-toastification"],
+  },
+});
